% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{auto_comment}
\alias{auto_comment}
\title{Add comments to database table from template}
\usage{
auto_comment(conn, db_table, timestamp = glue::glue("{today()} 09:00:00"))
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{db_table}{Either a dplyr connection to target table or a specification of 'schema.table'}

\item{timestamp}{Timestamp indicating when the comment is valid from}
}
\description{
Add comments to database table from template
}
