% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_store_helpers.R
\name{available_diseasystores}
\alias{available_diseasystores}
\title{Detect available diseasystores}
\usage{
available_diseasystores()
}
\value{
The installed diseasystores on the search path
}
\description{
Detect available diseasystores
}
\examples{
  available_diseasystores()  # DiseasystoreGoogleCovid19 + more from other packages
}
