% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{db_comment}
\alias{db_comment}
\title{Read or add a comment to the database}
\usage{
db_comment(
  conn,
  db_table,
  column = NA,
  comment = NULL,
  auto_generated = FALSE,
  timestamp = glue::glue("{today()} 09:00:00")
)

db_comment(
  conn,
  db_table,
  column = NA,
  comment = NULL,
  auto_generated = FALSE,
  timestamp = glue::glue("{today()} 09:00:00")
)
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{db_table}{Either a dplyr connection to target table or a specification of 'schema.table'}

\item{column}{The name of the table/column to get/set comment on (for table comment, use NA)}

\item{comment}{If different from NULL (default) the comment will added to the table/column. Use a reference "@column" too look up in docs.templates. Everything after the reference gets appended.}

\item{auto_generated}{Indicates if comment is auto generated or supplied by the user}

\item{timestamp}{Timestamp indicating when the comment is valid from}
}
\value{
The existing comment on the column/table if no comment is provided. Else NULL

The existing comment on the column/table if no comment is provided. Else NULL
}
\description{
Read or add a comment to the database

Read or add a comment to the database
}
