% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_digest_to_checksum_native_md5}
\alias{mg_digest_to_checksum_native_md5}
\alias{digest_internal}
\alias{mg_digest_to_checksum_internal}
\alias{mg_digest_to_checksum_internal.default}
\alias{mg_digest_to_checksum_internal.tbl_PqConnection}
\alias{mg_digest_to_checksum_internal.data.frame}
\alias{mg_digest_to_checksum_internal.tibble}
\title{Some backends have native md5 support, these use this function}
\usage{
mg_digest_to_checksum_native_md5(.data, col)

mg_digest_to_checksum_internal(.data, col)

\method{mg_digest_to_checksum_internal}{default}(.data, col)

\method{mg_digest_to_checksum_internal}{tbl_PqConnection}(.data, col)

\method{mg_digest_to_checksum_internal}{data.frame}(.data, col)

\method{mg_digest_to_checksum_internal}{tibble}(.data, col)

mg_digest_to_checksum_internal(.data, col)
}
\arguments{
\item{.data}{\cr
The data object to perform the operation on}

\item{col}{The name of column the checksums will be placed in}
}
\description{
Some backends have native md5 support, these use this function
}
