% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_R6_utils.R
\name{diseasyoption}
\alias{diseasyoption}
\title{Helper function to get option}
\usage{
diseasyoption(option, self)
}
\arguments{
\item{option}{(\code{character})\cr
Name of the option to get}

\item{self}{(\verb{R6::R6class Diseasy* instance})\cr
The object the option applies to.}
}
\description{
Helper function to get option
}
