% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_store_helpers.R
\name{diseasystore_exists}
\alias{diseasystore_exists}
\title{Check for the existence of a \code{diseasystore} for the case definition}
\usage{
diseasystore_exists(case_definition)
}
\arguments{
\item{case_definition}{(\code{character})\cr
A character string that controls which feature store to get data from.}
}
\description{
Check for the existence of a \code{diseasystore} for the case definition
}
