% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_featurestore.R
\name{drop_featurestore}
\alias{drop_featurestore}
\title{Drop feature stores from DB}
\usage{
drop_featurestore(
  pattern = ".*",
  schema = "ds",
  conn = diseasystore::mg_get_connection()
)
}
\arguments{
\item{pattern}{Pattern to match the tables by}

\item{schema}{Schema the diseasystore uses to store data in}

\item{conn}{DB connection}
}
\description{
Drop feature stores from DB
}
