% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_store_helpers.R
\name{get_diseasystore}
\alias{get_diseasystore}
\title{Get the \code{diseasystore} for the case definition}
\usage{
get_diseasystore(case_definition)
}
\arguments{
\item{case_definition}{case_definition (\code{character})\cr A character string that controls which feature store to get data from.}
}
\description{
Get the \code{diseasystore} for the case definition
}
