% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_log_error}
\alias{mg_log_error}
\alias{mg_log_warn}
\alias{log_printing}
\alias{mg_log_info}
\title{Convenience function to produce log output}
\usage{
mg_log_error(..., log_type = "ERROR")

mg_log_warn(..., log_type = "WARNING")

mg_log_info(..., tic = Sys.time(), log_file = "/dev/null", log_type = "INFO")
}
\arguments{
\item{...}{Arguments passed to paste (to produce log message)}

\item{log_type}{The type of information given to log (written before message in the log)}

\item{tic}{The timestamp used by the log entry (default Sys.time())}

\item{log_file}{Path to log-file. If not set, log entries are printed to console}
}
\description{
mg_log_error also calls stop(...). mg_log_warn also calls warning(...).
}
