% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_assert_id_like}
\alias{mg_assert_id_like}
\title{checkmate helper: Assert for "generic" mg_id structure}
\usage{
mg_assert_id_like(mg_id, ..., add = NULL)
}
\arguments{
\item{mg_id}{Object to test if is of class "Id" or character on form "schema.table"}

\item{...}{Parameters passed to checkmate::check_*}

\item{add}{\code{AssertCollection} to add assertions to}
}
\description{
checkmate helper: Assert for "generic" mg_id structure
}
