% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_create_logs_if_missing}
\alias{mg_create_logs_if_missing}
\title{Create a table with the mg log structure if it does not exists}
\usage{
mg_create_logs_if_missing(log_table, conn)
}
\arguments{
\item{log_table}{A specification of where the logs should exist ("schema.table")}

\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}
}
\description{
Create a table with the mg log structure if it does not exists
}
