% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_digest_to_checksum_native_md5}
\alias{mg_digest_to_checksum_native_md5}
\title{Some backends have native md5 support, these use this function}
\usage{
mg_digest_to_checksum_native_md5(.data, col)
}
\description{
Some backends have native md5 support, these use this function
}
