% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_filter_keys}
\alias{mg_filter_keys}
\title{Filters .data according to all records in the filter}
\usage{
mg_filter_keys(.data, filters, by = NULL, na_by = NULL)
}
\arguments{
\item{.data}{\cr
The data object to perform the operation on}

\item{filters}{A filter to subset data by.  if filters == NULL, no filtering occurs. Else, an mg_inner_join is performed using all columns of the filter}

\item{by}{passed to mg_inner_join if different from NULL}

\item{na_by}{passed to mg_inner_join if different from NULL}
}
\value{
An object similar to .data
}
\description{
If filter = NULL, not filtering is done
If filter is different from NULL, the .data is filtered by a mg_inner_join using all columns of the filter:
\code{mg_inner_join(.data, filter, by = colnames(filter))}

by and na_by can overwrite the mg_inner_join columns used in the filtering
}
