% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_format_timestamp}
\alias{mg_format_timestamp}
\title{formats timestamps as 2022-13-09 09:30,213}
\usage{
mg_format_timestamp(ts)
}
\arguments{
\item{ts}{Timestamp to format}
}
\description{
formats timestamps as 2022-13-09 09:30,213
}
