% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_get_schema}
\alias{mg_get_schema}
\title{Get the current schema of a DB connection}
\usage{
mg_get_schema(.x)
}
\arguments{
\item{.x}{A DBIConnection or lazy_query object}
}
\value{
The current schema name, but defaults to "prod" instead of "public"
}
\description{
Get the current schema of a DB connection
}
