% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_get_tables}
\alias{mg_get_tables}
\title{Gets the available tables}
\usage{
mg_get_tables(conn, pattern = NULL)
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{pattern}{A regex pattern with which to subset the returned tables}
}
\value{
A data.frame containing table names in the DB
}
\description{
Gets the available tables
}
