% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_in_schema}
\alias{mg_in_schema}
\title{Convenience function for dbplyr::in_schema}
\usage{
mg_in_schema(schema, table = NULL)
}
\arguments{
\item{schema}{Schema to look in / full specification for table "schema.table"}

\item{table}{Table to look for}
}
\description{
If schema is provided but table is not provided, schema is parsed assuming a format of "schema.table".
}
\seealso{
\link[dbplyr:in_schema]{dbplyr::in_schema} which this function wraps.
}
