% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_join_na_sql}
\alias{mg_join_na_sql}
\title{This function generates a much faster sql statement for NA join compared to dbplyr's _join with na_matches = "na".}
\usage{
mg_join_na_sql(by, na_by)
}
\description{
This function generates a much faster sql statement for NA join compared to dbplyr's _join with na_matches = "na".
}
