% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_join_warn_experimental}
\alias{mg_join_warn_experimental}
\title{A warning to users that SQL does not match on NA by default}
\usage{
mg_join_warn_experimental()
}
\description{
A warning to users that SQL does not match on NA by default
}
