% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_mg_age_labels}
\alias{mg_mg_age_labels}
\title{Provides age_labels that follows the mg standard}
\usage{
mg_mg_age_labels(age_cuts)
}
\arguments{
\item{age_cuts}{The lower bound of the groups (0 is implicitly included)}
}
\value{
A vector of labels with zero-padded numerics so they can be sorted easily
}
\description{
Provides age_labels that follows the mg standard
}
