% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_nrow}
\alias{mg_nrow}
\title{mg_nrow() but also works on remote tables}
\usage{
mg_nrow(.data)
}
\arguments{
\item{.data}{lazy_query to parse}
}
\value{
The number of records in the object
}
\description{
mg_nrow() but also works on remote tables
}
