% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_printr}
\alias{mg_printr}
\title{cat printing with default new line}
\usage{
mg_printr(..., file = "/dev/null", sep = "")
}
\arguments{
\item{...}{The normal input to cat}

\item{file}{Path of an output file to append the output to}

\item{sep}{The separator given to cat}
}
\description{
cat printing with default new line
}
