% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_reexports_db_timestamp}
\alias{mg_reexports_db_timestamp}
\alias{mg_db_timestamp}
\title{Determine the type of timestamps the DB supports}
\usage{
mg_db_timestamp(timestamp, conn)
}
\arguments{
\item{timestamp}{The timestamp to be transformed to the DB type. Can be character.}

\item{conn}{A \code{DBIConnection} to the DB where the timestamp should be stored}
}
\description{
Determine the type of timestamps the DB supports
}
