% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_slice_time}
\alias{mg_slice_time}
\title{Slices a data object based on time / date}
\usage{
mg_slice_time(.data, slice_ts, from_ts = from_ts, until_ts = until_ts)
}
\arguments{
\item{.data}{\cr
The data object to perform the operation on}

\item{slice_ts}{The time / date to slice by}

\item{from_ts}{The name of the column in .data specifying valid from time (note: must be unquoted)}

\item{until_ts}{The name of the column in .data specifying valid until time (note: must be unquoted)}
}
\value{
An object similar to .data
}
\description{
Slices a data object based on time / date
}
