% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_update_snapshot}
\alias{mg_update_snapshot}
\title{Update a historical table}
\usage{
mg_update_snapshot(
  .data,
  conn,
  db_table,
  timestamp,
  filters = NULL,
  message = NULL,
  tic = Sys.time(),
  log_file = TRUE,
  log_db = "prod.logs",
  enforce_chronological_order = TRUE
)

mg_update_snapshot(
  .data,
  conn,
  db_table,
  timestamp,
  filters = NULL,
  message = NULL,
  tic = Sys.time(),
  log_file = TRUE,
  log_db = "prod.logs",
  enforce_chronological_order = TRUE
)
}
\arguments{
\item{.data}{A tbl object. Typically a lazy query to the DB (from mg::get_table())}

\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{db_table}{Either a dplyr connection to target table or a specification of 'schema.table'}

\item{timestamp}{A timestamp (POSIXct) with which to update from_ts/until_ts columns}

\item{filters}{A filter to subset data by.  if filters == NULL, no filtering occurs. Else, an inner_join is performed using all columns of the filter}

\item{message}{A message to add to the log-file (useful for supplying metadata to the log)}

\item{tic}{A timestamp when computation began. If not supplied, it will be created at call-time. (Used to more accurately convey how long runtime of the update process has been)}

\item{log_file}{A logical that controls whether or not to produce a log file (useful for debugging). TRUE: log file is produced and copied to DB. FALSE: no log file and no updates to DB. NULL: no logfile but DB is updated}

\item{log_db}{The log table in the DB where logging information should be stored (default prod.logs)}

\item{enforce_chronological_order}{A logical that controls whether or not to check if timestamp of update is prior to timstamps in the DB}
}
\description{
Update a historical table

Update a historical table
}
\seealso{
mg_filter_keys

mg_filter_keys
}
