% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulist_data}
\alias{simulist_data}
\title{simulist_data}
\source{
Lambert J, Tamayo C (2024).
\emph{simulist: Simulate Disease Outbreak Line List and Contacts Data}.
\doi{10.5281/zenodo.10471458}, https://epiverse-trace.github.io/simulist/.
}
\description{
This data set contains a synthetic linelist created with the \code{simulist} package.
This linelist is used as an example data set for the \code{DiseasystoreSimulist} class.
}
\details{
The data set consists of a \code{tibble} with columns:
\itemize{
\item \code{id}: A unique identifier for each individual.
\item \code{case_type}: The type of case. One of "suspected", "probable", "confirmed".
\item \code{sex}: The sex of the individual.
\item \code{birth}: The birth date of the individual.
\item \code{age}: The age of the individual.
\item \code{date_onset}: The date of onset of symptoms.
\item \code{date_admission}: The date of admission to hospital.
\item \code{date_discharge}: The date of discharge from hospital.
\item \code{date_death}: The date of death.
}
}
\seealso{
\code{\link[simulist:simulist-package]{simulist::simulist()}}
}
\author{
Rasmus Skytte Randl\\u00F8v \email{rske@ssi.dk}
}
\keyword{data}
\keyword{internal}
