% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_interlace.R
\name{truncate_interlace}
\alias{truncate_interlace}
\title{Interlace the secondary tables by the keys and validity of the primary table}
\usage{
truncate_interlace(primary, secondary = NULL)
}
\arguments{
\item{primary}{The table with the keys the interlacing is done by (first "key_" column in primary).
The secondary tables are matched to the validity interval of the primary table before interlacing}

\item{secondary}{A list of tables that should be matched and joined to the primary table}
}
\description{
A short description...
}
