% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{aggregators}
\alias{aggregators}
\alias{key_join_sum}
\alias{key_join_max}
\alias{key_join_min}
\alias{key_join_count}
\title{Feature aggregators}
\usage{
key_join_sum(.data, feature)

key_join_max(.data, feature)

key_join_min(.data, feature)

key_join_count(.data, feature)
}
\arguments{
\item{.data}{.data\cr The data object to perform the operation on}

\item{feature}{(\code{character})\cr
Name of the feature to perform the aggregation over}
}
\description{
Feature aggregators
}
