% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasystore-package.R
\docType{package}
\name{diseasystore-package}
\alias{diseasystore}
\alias{diseasystore-package}
\title{diseasystore: feature store handler for the diseasy framework}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package handles the feature store data model for the diseasy framework.
}
\author{
\strong{Maintainer}: Rasmus Skytte Randløv \email{rske@ssi.dk} (\href{https://orcid.org/0000-0002-5860-3838}{ORCID})

}
\keyword{internal}
