% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_store_helpers.R
\name{diseasystore_case_definition}
\alias{diseasystore_case_definition}
\title{Transform case definition to PascalCase}
\usage{
diseasystore_case_definition(case_definition)
}
\arguments{
\item{case_definition}{case_definition (\code{character})\cr A character string that controls which feature store to get data from.}
}
\value{
The given case_definition formatted to match a Diseasystore
}
\description{
Transform case definition to PascalCase
}
\examples{
  diseasystore_case_definition("Google COVID-19")  # DiseasystoreGoogleCovid19
}
