% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasystoreBase.R
\name{DiseasystoreBase}
\alias{DiseasystoreBase}
\title{diseasystore base handler}
\description{
This \code{DiseasystoreBase} \link[R6:R6Class]{R6} class forms the basis of all feature stores.
It defines the primary methods of each feature stores as well as all of the public methods.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fs_map}}{(\verb{named list}(\code{character}))\cr
A list that maps features known by the feature store to the corresponding feature handlers
that compute the features.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasystoreBase-new}{\code{DiseasystoreBase$new()}}
\item \href{#method-DiseasystoreBase-get_feature}{\code{DiseasystoreBase$get_feature()}}
\item \href{#method-DiseasystoreBase-key_join_features}{\code{DiseasystoreBase$key_join_features()}}
\item \href{#method-DiseasystoreBase-clone}{\code{DiseasystoreBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreBase-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasystoreBase} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreBase$new(
  start_date = NULL,
  end_date = NULL,
  slice_ts = NULL,
  source_conn = NULL,
  target_conn = NULL,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{(\code{Date})\cr
Study period start.}

\item{\code{start_date}}{(\code{Date})\cr
Study period start.}

\item{\code{end_date}}{(\code{Date})\cr
Study period end.}

\item{\code{end_date}}{(\code{Date})\cr
Study period end.}

\item{\code{slice_ts}}{(\code{Date} or \code{character})\cr
Date to slice the database on (used if source_conn is a database)}

\item{\code{slice_ts}}{(\code{Date} or \code{character})\cr
Date to slice the database on (used if source_conn is a database)}

\item{\code{source_conn}}{\cr
Used to specify where data is located.
Can be \code{DBIConnection} or file path depending on the \code{diseasystore}.}

\item{\code{target_conn}}{(\code{DBIConnection})\cr
A database connection to store the computed features in.}

\item{\code{verbose}}{(\code{boolean})\cr
Boolean that controls enables debugging information.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the \code{DiseasystoreBase} \link[R6:R6Class]{R6} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreBase-get_feature"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreBase-get_feature}{}}}
\subsection{Method \code{get_feature()}}{
Computes, stores, and returns the requested feature for the study period.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreBase$get_feature(
  feature,
  start_date = private \%.\% start_date,
  end_date = private \%.\% end_date,
  slice_ts = private \%.\% slice_ts
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{(\code{character})\cr
The name of a feature defined in the feature store.}

\item{\code{start_date}}{(\code{Date})\cr
Study period start.}

\item{\code{start_date}}{(\code{Date})\cr
Study period start.}

\item{\code{end_date}}{(\code{Date})\cr
Study period end.}

\item{\code{end_date}}{(\code{Date})\cr
Study period end.}

\item{\code{slice_ts}}{(\code{Date} or \code{character})\cr
Date to slice the database on (used if source_conn is a database)}

\item{\code{slice_ts}}{(\code{Date} or \code{character})\cr
Date to slice the database on (used if source_conn is a database)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tbl_dbi with the requested feature for the study period.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreBase-key_join_features"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreBase-key_join_features}{}}}
\subsection{Method \code{key_join_features()}}{
Joins various features from feature store assuming a primary feature (observable)
that contains keys to witch the secondary features (defined by aggregation) can be joined.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreBase$key_join_features(observable, aggregation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observable}}{(\code{character})\cr
The name of a feature defined in the feature store}

\item{\code{aggregation}}{(\code{list}(\code{quosures}))\cr
Expressions in aggregation evaluated to find appropriate features.
These are then joined to the observable feature before aggregation is performed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tbl_dbi with the requested joined features for the study period.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
