% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasystoreGoogleCovid19.R
\name{DiseasystoreGoogleCovid19}
\alias{DiseasystoreGoogleCovid19}
\title{feature store handler of Google Health COVID-19 Open Data features}
\value{
A new instance of the \code{DiseasystoreGoogleCovid19} \link[R6:R6Class]{R6} class.
}
\description{
This \code{DiseasystoreGoogleCovid19} \link[R6:R6Class]{R6} brings support for using the Google
Health COVID-19 Open Data repository.
See the vignette("google_covid_19_data") for details on how to configure the feature store
}
\examples{
  ds <- DiseasystoreGoogleCovid19$new(source_conn = ".",
                                      target_conn = DBI::dbConnect(RSQLite::SQLite()))

}
\section{Super class}{
\code{\link[diseasystore:DiseasystoreBase]{diseasystore::DiseasystoreBase}} -> \code{DiseasystoreGoogleCovid19}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasystoreGoogleCovid19-key_join_filter}{\code{DiseasystoreGoogleCovid19$key_join_filter()}}
\item \href{#method-DiseasystoreGoogleCovid19-clone}{\code{DiseasystoreGoogleCovid19$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="finalize"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-finalize'><code>diseasystore::DiseasystoreBase$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="get_feature"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-get_feature'><code>diseasystore::DiseasystoreBase$get_feature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="initialize"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-initialize'><code>diseasystore::DiseasystoreBase$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="key_join_features"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-key_join_features'><code>diseasystore::DiseasystoreBase$key_join_features()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreGoogleCovid19-key_join_filter"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreGoogleCovid19-key_join_filter}{}}}
\subsection{Method \code{key_join_filter()}}{
This function implements an intermediate filtering in the aggregation pipeline.
For semi-aggregated data like Googles COVID-19 data, some people are counted more than once.
The \code{key_join_filter} is inserted into the aggregation pipeline to remove this double counting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreGoogleCovid19$key_join_filter(
  .data,
  aggregation_features,
  start_date = private \%.\% start_date,
  end_date = private \%.\% end_date
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{(\code{any})\cr The data object to perform the operation on.}

\item{\code{aggregation_features}}{(\code{character})\cr
A list of the features included in the aggregation process.}

\item{\code{start_date}}{(\code{Date})\cr Study period start.}

\item{\code{end_date}}{(\code{Date})\cr Study period end.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A subset of \code{.data} filtered to remove double counting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreGoogleCovid19-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreGoogleCovid19-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreGoogleCovid19$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
