% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasystoreSimulist.R
\name{DiseasystoreSimulist}
\alias{DiseasystoreSimulist}
\title{feature store handler of synthetic \code{simulist} features}
\value{
A new instance of the \code{DiseasystoreSimulist} \link[R6:R6Class]{R6} class.
}
\description{
This \code{DiseasystoreSimulist} \link[R6:R6Class]{R6} brings support for individual level data.
}
\examples{
  ds <- DiseasystoreSimulist$new(
    source_conn = ".",
    target_conn = DBI::dbConnect(RSQLite::SQLite())
  )

  rm(ds)
}
\section{Super class}{
\code{\link[diseasystore:DiseasystoreBase]{diseasystore::DiseasystoreBase}} -> \code{DiseasystoreSimulist}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasystoreSimulist-new}{\code{DiseasystoreSimulist$new()}}
\item \href{#method-DiseasystoreSimulist-clone}{\code{DiseasystoreSimulist$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="finalize"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-finalize'><code>diseasystore::DiseasystoreBase$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="get_feature"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-get_feature'><code>diseasystore::DiseasystoreBase$get_feature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasystore" data-topic="DiseasystoreBase" data-id="key_join_features"><a href='../../diseasystore/html/DiseasystoreBase.html#method-DiseasystoreBase-key_join_features'><code>diseasystore::DiseasystoreBase$key_join_features()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreSimulist-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreSimulist-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasystoreSimulist} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreSimulist$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to the \code{?DiseasystoreBase} constructor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the \code{DiseasystoreSimulist} \link[R6:R6Class]{R6} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasystoreSimulist-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasystoreSimulist-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasystoreSimulist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
