% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_diseasystore.R
\name{drop_diseasystore}
\alias{drop_diseasystore}
\title{Drop feature stores from DB}
\usage{
drop_diseasystore(
  pattern = NULL,
  schema = diseasyoption("target_schema"),
  conn = SCDB::get_connection()
)
}
\arguments{
\item{pattern}{Pattern to match the tables by}

\item{schema}{Schema the diseasystore uses to store data in}

\item{conn}{DB connection}
}
\value{
No return value, called for side effects
}
\description{
Drop feature stores from DB
}
\examples{
  conn <- SCDB::get_connection(drv = RSQLite::SQLite())

  drop_diseasystore(conn = conn)

  DBI::dbDisconnect(conn)
}
