% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_assert_dbtable_like}
\alias{mg_assert_dbtable_like}
\title{checkmate helper: Check for "generic" db_table type}
\usage{
mg_assert_dbtable_like(db_table, ...)
}
\arguments{
\item{db_table}{Object to test if is of class "tbl_dbi" or character on form "schema.table"}

\item{...}{parameters passed to checkmate::check_*}
}
\description{
checkmate helper: Check for "generic" db_table type
}
