% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_db_timestamp}
\alias{mg_db_timestamp}
\title{Determine the type of timestamps the DB supports}
\usage{
mg_db_timestamp(timestamp, conn)

mg_db_timestamp(timestamp, conn)
}
\description{
Determine the type of timestamps the DB supports

Determine the type of timestamps the DB supports
}
