% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_digest_to_checksum}
\alias{mg_digest_to_checksum}
\title{Computes an MD5 checksum from columns}
\usage{
mg_digest_to_checksum(.data, col = "checksum", warn = TRUE)
}
\arguments{
\item{.data}{A tbl object. Typically a lazy query to the DB (from mg::get_table())}

\item{col}{Name of the column to put the checksums in}

\item{warn}{Flag to warn if target column already exists in data}
}
\description{
Computes an MD5 checksum from columns
}
