% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_id}
\alias{mg_id}
\title{Convenience function for DBI::Id}
\usage{
mg_id(db_table_name, conn = NULL)
}
\arguments{
\item{db_table_name}{Name of the table to return. Can implicitly contain information about schema.}

\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}
}
\description{
Convenience function for DBI::Id
}
\seealso{
\link[DBI:Id]{DBI::Id} which this function wraps.
}
