% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_is.historical}
\alias{mg_is.historical}
\title{Checks if lazy_query query contains historical data}
\usage{
mg_is.historical(.data)
}
\arguments{
\item{.data}{\cr
The data object to perform the operation on}
}
\value{
TRUE if .data contains the columns: "checksum", "from_ts", and "until_ts". FALSE otherwise
}
\description{
Checks if lazy_query query contains historical data
}
