% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_reexports.R
\name{mg_table_exists}
\alias{mg_table_exists}
\title{Check if table exists in db}
\usage{
mg_table_exists(conn, db_table_id)
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{db_table_id}{A \code{\link[DBI:Id]{DBI::Id()}} object or a character string readable by \link{mg_id}.}
}
\description{
Check if table exists in db
}
