% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_nolint_linter.R
\name{nolint_linter}
\alias{nolint_linter}
\title{nolint linter}
\usage{
nolint_linter(length = 80L)
}
\arguments{
\item{length}{maximum line length allowed. Default is 80L (Hollerith limit).}
}
\description{
Check that the \verb{nolint:} statements occur after the character limit
}
\examples{
# will produce lints
lint(
  text = paste(strrep("x", 15L), "# nolint: nolint_linter",
  linters = nolint_linter(length = 20L)
)

# okay
lint(
  text = paste(strrep("x", 19L), "# nolint: nolint_linter",
  linters = nolint_linter(length = 20L)
)

}
\seealso{
\itemize{
\item \link{linters} for a complete list of linters available in lintr.
\item \url{https://style.tidyverse.org/syntax.html#long-lines}
}
}
